projects/bank-analyzer/pyproject.toml
[project]
name = "bank-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "babel>=2.18.0",
    "google-genai>=1.71.0",
    "joblib>=1.5.3",
    "ofxparse>=0.21",
    "pandas>=3.0.2",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "quiffen>=4.0.1",
    "reportlab>=4.4.10",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
]
